add didchangeWatchedFiles in lsp msg broker queue
